pdf_extraction_service/requirements.txt
# Core Dependencies
fastapi>=0.109.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.5.0
google-genai>=0.2.0
PyPDF2>=3.0.0
pdf2image>=1.16.3
pytesseract>=0.3.10
pika>=1.3.0
python-multipart>=0.0.6
python-magic>=0.4.27
httpx>=0.26.0

# Image Processing
Pillow>=10.0.0
numpy>=1.24.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0 